Refactor schema package build to use tsc
