+++ packages/schema/src/scripts/data/board.json
