+++ packages/schema/src/scripts/update-json-schema.ts
@@ -6,12 +6,12 @@
 
 import fs from "fs";
 import path from "path";
-import packageJson from "../package.json";
+import packageJson from "../../package.json" assert { type: "json" };
 
 const PACKAGE_ROOT = process.cwd();
 const SCHEMA_PATH = path.relative(PACKAGE_ROOT, "breadboard.schema.json");
 
-const schema = JSON.parse(fs.readFileSync(SCHEMA_PATH, "utf8"))
+const schema = JSON.parse(fs.readFileSync(SCHEMA_PATH, "utf8"));
 
 console.log(`current schema id: ${schema.$id}`);
 console.log(`package.json version: ${packageJson.version}`);
@@ -20,7 +20,7 @@ const GITHUB_OWNER = "breadboard-ai";
 const GITHUB_REPO = "breadboard";
 const GITHUB_REF = `@google-labs/breadboard-schema@${packageJson.version}`;
 
-const PACKAGE_PATH = "packages/schema"
+const PACKAGE_PATH = "packages/schema";
 
 schema.$id = `https://raw.githubusercontent.com/${GITHUB_OWNER}/${GITHUB_REPO}/${GITHUB_REF}/${PACKAGE_PATH}/${SCHEMA_PATH}`;
 console.log(`updating schema id to: ${schema.$id}`);
